Adds "maybe" parser
Adds a "maybe" parser combinator. This is just an alias for something
like `oneOf(parser, success(default))`. In the future, I may add a
`Maybe` type, so a default is not required. However this would make
handling a bit more awkward, especially with `map`. So I went with this
simpler option for now.

Updates RegularLanguages example to use `maybe`.

Cleans up `oneOf` definition. No change in behavior.