Adds `lazy` combinator allowing for recursion!
Adds the lazy combinator which allows for recursion, and thus parsing
of context free languages. This is a relatively small change for a huge
gain in parsing abilities.

Adds a new examples file to demonstrate that we are indeed able to parse
languages which are context free and not regular.