Updates to index-based parsing
Updates parsers to use an index to track the progress of parsing. Until
now, parsers would return the rest of the string to be parsed. Instead,
parsers now do not modify the input, and instead update the current
index that the next parser should start from.

This should be a little more efficient, and should allow for better
error messages in the future.