Updates parsers to use new error message format
Updates the existing parser to use the improved error message format.

Updates tests which check error messages.

Minor change to the dateParser example. I found an interesting effect
in the date parser which I hadn't noticed before. Since `oneOf` tries
parsers in order, the section
```
oneOf(
    sequence(exact("-"), ..., exact("-"), ...),
    sequence(exact("/"), ..., exact("/"), ...),
);
```
when parsing "-01/01" will fail on the first sequence at "/", then try
the second sequence, which will fail at "/". So the error will occur at
the first "-". This makes it appears as though the date parser has a
preference between separators. It would be interesting to try and
implement a `commit` combinator which somehow indicates to `oneOf` that
it shouldn't go on to the next option and instead fail on this one.
I think the Elm parser has something this but reverse. They have a
"backtrackable" combinator.

Updates example in Readme.