Adds basic Pratt parser
Adds a basic Pratt parser. Currently this is limited to binary infix
operators. New operators can be defined with a pair of numbers
representing the "binding power" of the operator. This pair determines
the precedence of the operator and whether it should associate to the
left or right.

Adds an example which demonstrates this and how it compares to the
"standard" BNF style approach which produces only right associating
parsers.