Generalizes Pratt parser for better operator types
Adds more type parameters to generalize types for operators. This means
operators can have types attached to them and are no longer required to
be strings.

Updates terminology from "map" to "acc" for 'accumulator'. This is a
slightly more accurate description of how this function is meant to be
used.